requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
python-docx==0.8.11
joblib==1.3.2
plotly==5.16.1
dash==2.13.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.2
wtforms==3.0.1
flask-wtf==1.1.1
pytest==7.4.2
black==23.7.0
flake8==6.1.0
